Refactor: Draw Batches with a Single Material
In previous versions, the Renderer had the ability to swap materials as
draw operations came in. In this commit, that behaviour is replaced with
one where only one material and texture is used for all drawing
operations. Draw operations (ex. `Renderer::draw_rect`) can also
be individually configured with different palettes, by passing in a
palette atlas and the desired index.

Changes:
 * Add `palette_atlas` and `palette_index` properties to SpriteComponent
 * Instead of taking in a material, let `Renderer::draw_rect` take in a
   palette atlas and index
 * Fix bug that prevented `Material::set_texture` from correctly
   overwriting a previously assigned texture
 * Replace various reference types with const references