bld: Compile Engine and Game Separately
In this commit, the project is split into two subdirectories: `engine`
and `mco`. `engine` represents the engine and compiles into a library.
`mco` represents the actual game which links to the engine and compiles
into an `executable`.

This may not be an ideal setup and will likely be changed in the future.