Bld: Separate Build Tree from Source
This commit restructures the build system to be much more clean and
flexible. Previous versions split the cmake project into an engine
subproject and a game subproject, and the build tree would be split into
three. This was difficult to work with, but this new approach solves
many of the prior issues. All build files exist at the top level build
directory of the project, and the project is partitioned into four major
targets: `engine`, `base`, `client`, and `server`. `CMakeList.txt` files
are also much shorter and concise compared to earlier commits.